pyproject.toml
[project]
name = "eduai-tutor"
version = "0.1.0"
description = ""
authors = [
    { name = "Dorothée Catry", email = "catrydorothee@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"  # ✅ Limite haute pour django-tailwind

dependencies = [
    "django >=5.2.4,<6.0.0",
    "djangorestframework >=3.16.0,<4.0.0",
    "django-tailwind[reload] (>=4.0.1,<5.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-widget-tweaks (>=1.5.0,<2.0.0)",
    "chromadb (>=1.0.15,<2.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pypdf (>=5.7.0,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "ollama (>=0.5.1,<0.6.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
